Allow to disable actions on a per-row (per-record) basis
When adding actions to a dataGrid the `condition` key can be specified. It can hold a function which currently doesn't receive any parameters.

When we pass the record here we allow developers to decide whether or not to show a action on a per-row (per-record) basis, so showing an action for some items, hiding it for others.